Stop truncating ex-MyISAM tables in tests
We were doing this on MySQL because the hits tables were using MyISAM,
which doesn't support transactions (which are the default strategy for
database_cleaner). There shouldn't be any reason to do this on PostgreSQL,
so let's stop.

For some tests which have a lot of setup, we have been using testing_before_all
to indicate that their test data should be cleaned up only once at
the end, and with truncation, to speed up test runs. This commit assumes that:

- that's still a good thing to do for those cases;
- we have always been consistent about using truncate_everything
  for only MyISAM reasons, and testing_before_all for tests with lots
  of data.