Convert FOI response script into a rake task
This means that Postgres authentication will work as normal, and the
task can be run from the root of the project.

There are some mappings rows which cause encoding problems such as
'Encoding::UndefinedConversionError: "\xE2" from ASCII-8BIT to UTF-8'.
The paths do look somewhat mangled, so this catches the exception and
excludes those rows for now. 616 rows in my development database are
excluded by this.

This still doesn't have any tests.