Remove code for manually updating organisation slugs
This will be done automatically whenever organisations are imported.
Organisations are currently imported before site config is imported.

The code that has been removed also updated users. This is no longer necessary
as the organisation_slug stored against user isn't used for anything, and will
be removed in a future of gds-sso.

It will still be necessary to update any organisation slug changes in
transition-config.