No longer create/support fake organisations
The Directgov and Businesslink organisations were a fudge to provide a separate
list of sites.

Now that all of the sites associated with these fake organisations have been
moved to other organisations, we can remove this complexity.

By doing this, we make the code simpler and we make it easier to switch to
using content IDs as the primary key for organisations.