Fix generation of links to remove individual filters
The links to remove individual filters from the mappings index pages
were being wrongly generated by the site_mappings_path helper using
the site_id from the params hash as the format, for example:

    /sites/aaib/mappings.aaib?type=redirect

Visiting those links raised ActionController::UnknownFormat, so 3 features
were failing.

This started happening between Rails 4.1.9 and 4.1.10, and it seems to be
related to changes to handling of positional arguments which were backported
to 4.1.10 [1]. Those changes were made by @evilstreak and @tekin to fix
a bug they encountered in Whitehall [2] and the comment there explains
more of the context.

In this case we were passing the site object to the helper method as well as
site_id appearing in params, and the latter was being interpreted as the
format. To work around this we're no longer passing the site object, only
the params.

[1]: https://github.com/rails/rails/pull/18803
[2]: https://github.com/alphagov/whitehall/commit/5d46001d7846a85180327797f95c72abbeac3b5e