Break shared_examples up and reduce hard-coding
'creates mappings' still looks like a reasonable use of shared_examples.
Unfortunately its consumers all use identical example data and that's
hard-coded into the example.

I'm planning to add another test covering processing a batch that includes an
archive mapping with a custom archive URL. Not all of the shared_examples'
consumers will want to do that, so it should exist outside of the main example.

In preparation for that, for consistency I've moved the test for processing
a redirect mapping to its own shared example. This and my new, upcoming archive
test will sit side-by-side.

- It wasn't clear from the existing specs + shared_examples how the tag_list
  should be tested - I assume it should be provided as a comma-separated list
  and should be returned as an array, but might be wrong.

- There's a nonsense `context` in the new shared example block. It _seemed_ to
  be necessary to have a context for the setup/teardown to function properly
  and not interfere with other shared_examples.