Allow access to CSV's custom archive URLs
We currently only make use of a CSV row's `new_url` column when the row
represents a redirect mapping. That redirect's new_url then goes on to
eventually be the `new_url` for the persisted mapping record.

Meanwhile, a CSV row representing an archive mapping has its new_url ignored.

This change is the first step in the process of getting a custom archive URL
from the CSV's new_url column and assigning it to the resulting archive
mapping's archive_url attribute.