Assign CSV row's archive_url to its entry in batch
In the early stages of a CSV import, we create an ImportBatch record which has
many 'import batch entries'.

For each row in the CSV, we create a batch entry and assign to it the CSV row's
values.

This change includes `archive_url` (a custom archive URL) in that group of
values.

By this point in the CSV import, ImportBatchRow has already validated that the
archive_urls are non-blank and match
`NationalArchivesURLValidator::NATIONAL_ARCHIVES_HOST`, so in ImportBatch we
only validate that these URLs aren't too long.