Define precedence for duplicate archive mappings
When the same Old URL is listed to be archived multiple times in the same CSV
for an import, we ignore the later entries in favour of the first one found.

This commit does not affect how archived mappings are compared to archived
mappings with custom New URLs. Archived mappings are treat the same whether
they have a custom URL or not.