Define precedence for custom archive mappings
This commit makes archives with custom New URLs take precendence over archives
with no custom URL specified.

Just like we do with regular (non-custom URL) archive mappings, when the same
Old URL is listed to be archived multiple times in the same CSV for an import,
we ignore the later entries in favour of the first one found.

For example, when importing the following CSV of 2 regular archives and
2 custom archives:

  Old URL,New URL
  http://foo.gov.uk/a,TNA
  http://foo.gov.uk/a,TNA
  http://foo.gov.uk/a,http://webarchive.nationalarchives.gov.uk/20120816224015/http://foo.org.uk/a
  http://foo.gov.uk/a,http://webarchive.nationalarchives.gov.uk/20120816224015/http://foo.org.uk/b

The Old URL http://foo.gov.uk/a will be mapped to the New URL
  http://webarchive.nationalarchives.gov.uk/20120816224015/http://foo.org.uk/a