Add a rake task to revert import of some sites
We occasionally configure sites and then later decide that they should not be
transitioned via the Transition tool. In these cases it's good to delete the
imported site so that the data reflects our decisions. An example of such a
site is the content planner application on alphagov.co.uk: this app is no
longer in Production before the move to publishing.service.gov.uk so we don't
need to redirect the old domain, but it has already been imported.

This rake task makes it easier to delete these sites and all associated data
which may have been created by the site import: hosts and links to extra
organisations. The site should be removed from the transition-config repo
before running the task, otherwise the site will be re-imported the next time
the orgs_sites_hosts import runs.

The task takes an array of site abbrs, eg:

    bundle exec rake import:revert:sites[gds_alphagov_content-planner,another_site]

If any given site has any mappings or hits rows, it will not be deleted. The
presence of these is strongly indicative of us needing to keep the sites in
the database; if we really do want to delete a site which has mappings or hits
rows then that is a very special case and won't be automated.