Use correct domain for Whitehall mappings import
We're no longer using the alphagov.co.uk domain for our applications, so
should get the correct domain from the environment via Plek. Since the
migration to the *.publishing.service.gov.uk domains this import has been
erroring with `OpenURI::HTTPError: 401 Unauthorized` - I suspect the basic
auth credentials aren't being passed along when following the redirect.

This changes the behaviour to use Whitehall in the current environment rather
than always using Production, but there is no mention in git history of any
reason for doing that. It's more helpful to test against the current
environment so that any errors due to changes in Whitehall can be detected
before getting to Production.