Remove warning about "empty" raise_error matcher
The "empty" raise_error matcher will catch errors that might hide bugs in
our code such as NoMethodError so we should prefer a raise_error with an
argument to specify something to match any exceptions against.  In this
case the mock was configured incorrectly to return a raise_error matcher
instead of actually raising an error.  Once that is fixed we can match
against the error message in the raised error.