Remove warning about "should"-style as_null_object
Calling .as_null_object on any existing object is deprecated along with
the old "should" style rspec syntax.  Instead we construct an explict
double object, call .as_null_object on it and provide it to the object
under test.  We have to use `instance_variable_set` here because neither
the Precompute nor the ConsoleJobWrapper provide a setter or initialize
argument for the console.

A better solution would be to refactor to allow this.