Fix failures introduced by transpec and old `have` matchers
Rspec used to allow specs of the form `subject.should have(n).things`.
The `things` in this could mean two different things.  First it could be a
pointless modifier to help the spec read better (e.g. if subject is an
array of names we could write `list.should have(5).names` or
`array_of_names.should have(3).people_in_it` and this would be fine). The
other meaning it takes is that of a reference to a method call on our
subject to assert that the result of the method call has that many items
in it.  (e.g. if subject was an activerecord model with a hits method we
could write `site.should have(2).sites` and this would check that the sites method returns something that size of 2).  Transpec can't tell which meaning
we have used so it has replaced some with `expect(subject.things.size).to eq(n)` and others with `expect(subject.size).to eq(n)`.

This commit fixes the ones it got wrong.  This is made more confusing
because in the hits_controller_spec the subject was a Struct.new instance
and these have a size method that returns the number of fields in the
struct, allowing the `subject.size` to pass without a NoMethodError, but
produce the wrong result (in this case we wanted `subject.points.size`).