Remove ci_reporter gem
We're not on the most recent version and previous experience
suggests upgrading to make it work for both cukes and specs is not
simple.  Happily, the report junit report task isn't enabled on
jenkins (and doesn't appear to have been for a while) so it's easier
to remove the gem entirely.