Bump versions of some rails integration gems
activerecord-session_store, kaminari, whenever, gretel, and
acts-as-taggable-on

The upgrades are straightforward minor or patch level upgrades. The
acts-as-taggable-on upgrade brings a couple of migrations although one
is a no-op because we don't use mysql.  It also brings in a change
that we wanted which is to strip quotes from tags if the tag doesn't
need it (e.g. a tag `"hello"` would become just `hello`).  We had tests
that ensured it didn't do this (e.g. the tag would remain `"hello"`), but
checking on integration it seems none of our tags are this shape so
there's no problem accepting this new behaviour and changing the tests
- no data will be lost or inconsistent.