Bump activerecord-import
Our version was fairly out of date and the changelog on the project is to
trawl the commits.  As it's only used in one place (BulkAddBatch.create) I
was considering removing the dependency - but it turns out it's important.
Creating a BulkAddBatch might involve creating 1000s of BulkAddBatchEntry
rows and this needs to happen in the request/response rather than being
farmed off to a background worker like other high-volume object creating/
editing processes.  Luckily there's reasonable test coverage and the
version bump was smooth.