Bump to rails 4.2
This includes using railsdiff to compare a fresh 3.2.18 app to a fresh
4.2.6 one and incorporating those differences to our app.  This commit
only upgrades from 4.1 to 4.2, but the process hadn't been run since the
initial 3.2 app was created so we still had some out of date config
and we doing things the old way (secrets.yml for example).

Note: at this point specs do not run because we have to deal with some
special things that this app does which are no longer easily supported
in rails 4.2.