Remove the disable_columns concern
In Rails 4.2 attributes is no longer a simple Hash object so this concern
no longer works.  I attempted to rewrite the concern to work with the new
object (ActiveRecord::AttributeSet) but it's got a much smaller api
surface than a Hash, and as such is much less mutable.  This is probably
a good thing from a library api design, but it makes attempting to hide
db columns from the model much harder.

I tried reverting to the old soluiton of filtering the columns from the
db, but although that meant `type` and `new_url` were no longer in the
attributes for an ImportBatch, they were still defined as attribute
methods on the class (it gets them from the base class which still
includes those columns).  Next approach was to try undef-ing the methods
but I couldn't get it to work without blowing other things up.

In the end I decided the pragmatic thing to do is simply remove the
concern.  Yes an ImportBatch shouldn't really have a type or new_url
attribute, but it's fighting against ActiveRecord to make this happen.