Tell cucumber to allow rails to use custom exceptions app
The intention of the `@allow-rescue` tag is that we allow exceptions to be
rescued so we can assert that our custom exceptions are working.  For most
rails apps this is a rescue_from in a controller, for us however it's
handled by our exceptions_app and the settings that `@allow-rescue` sets
are not enough.  We have to disable the DebugExceptions middleware which
sits below the ShowExceptions middleware (ultimately controlled by the
consider_all_requests_local setting) and then ask ShowExceptions to call
out to our exception_app (via the show_exceptions setting).

This is slightly more complex than expected because these settings
accessed in two ways: via the config setting directly, but also via the
rack environment.  The rack environment is constructed from the config
settings, but it is memoized after first request so we need to mutate
that environment as well as changing the settings.