Change yaml fixture for whitehall orgs to hashes, not openstructs
There is a bug in ruby 2.3.0 (https://bugs.ruby-lang.org/issues/11884)
that means serializing/deserializing openstructs from yaml is broken. We
can avoid it entirely by converting the yaml file to contain simple hashes
and then use the GdsApi::Response.build_openstruct_recursively method to
turn the hashes into the openstructs we expect.  It might be a pointer
that this object should not use a magic attribute that is not set in live
to change behaviour, but instead we should stub the response using the
gdsapi test helpers.  But this'll do for now.