Add note about our config.exceptions_app = self.routes
As of rails 4.2 it's no longer mentioned in the guide as something to
consider doing to get dynamic error pages.  See: https://github.com/rails/rails/pull/17815
Setting an exceptions_app is still documented, but setting it to
self.routes is not.  It still works for us, but we should be aware that
this might be a blocker to future upgrades.