Make ImportedHitsFile robust when filename doesn't exist
It'll still fail validation because the content_hash won't be set, but it
won't fail by raising an Errno::ENOENT.  We make this change because
shoulda-matchers now care about case-sensitivity for
validates_uniqueness_of and the test it creates for that uses a
filename that doesn't exist (e.g. the upcased version of the one that
does).