Fix black lines on hits graphs
Between 28/04/2016 and 03/05/2016 the hits graphs suddenly gained a black line
for every day, instead of only for the transition date. I can't see any
relevant changes in the Google Charts release notes [1] - the last documented
release was in February, long before we started seeing these problems.

The documentation for chart annotations [2] says that the default value for
the annotation role is '', which suggests that our existing code should still
work, but the examples on that page use null for datapoints without
annotations. Switching our default value in the helper to nil appears to fix
the problem, and the annotation for the transition date is shown correctly.

[1]: https://developers.google.com/chart/interactive/docs/release_notes#official-releases
[2]: https://developers.google.com/chart/interactive/docs/roles#what-roles-are-available