Import hits with .jsp extension
We already ignore hits rows which contain various patterns which indicate that
we don't need to map them, such as `.js`, but we do want to import hits rows
with .jsp - this is an extension used by Java Spring apps.

This commit updates the pattern to match either .js$ or .js<any non-word
character> so that any .js paths with query strings are still matched. This
enables us to import .jsp hits.

I don't intend to reimport all the hits files after this change has been
deployed. There's one site in particular which I know is affected by this
(directgov_local) so we will reimport those hits files from
pre-transition-stats.