Update poltergeist and make phantomjs an explicit dependency
We have failing tests when we try to interact with the charts on the hits summary pages.  Investigation shows that the JS to draw the chart is not being triggered by the google charts API we use.  As this is a JS problem and our JS testing infrastructure is a little out of date, our first step is to upgrade these dependencies.

We also configure poltergeist to use the binary that the phantomjs gem
will install rather than relying on whatever version is installed on the
dev or ci machines.  This means we get a newer version of phantomjs that
fixes many bugs and is more like a real browser.