Use a larger poltergeist window to avoid graph label truncation
Upgrading poltergeist, phantomjs, and the chart api loading code was
enough to make the charts render in the tests again.  However, they
were still failing when trying to check the text of the labels in the
chart legend.  The text of the "Redirects" label was being truncated
to "Redire…" which is something the charts api does if the label area
is too small.  This isn't something that happens in the browser, nor
is it something that happens in the browser if we make it as small as
the poltergeist window (1024x768).  However, if we make the poltergeist
window wider (1366x768 - the default for maximising the window) it
works does not truncate the labels and the tests passs again.