Ignore malformed URIs
Some of the hits data has malformed URIs, for example

/statistics/child%E2?%3Fpoverty%E2%3F%3Fstats.htm

instead of

/statistics/child-poverty-stats.htm

This is preventing us from running the transition import:
the Transition_load_all_data jenkins job fails as soon as it tries
to handle one of these paths.

There is already a HostPath entry for the correct path, as well
as the one with the malformed the URI.

This change means we skip associating the HostPath to the Mapping
if the URI cannot be decoded.