Fix tag list field to fill with comma separated tags
Previously the text_field for a tag_list_field with multiple tags would be filled with words only separated by a space (e.g. "fee fi of"). The javascript would turn these into concatenated tags and save them as a new tag on save. The problem seems to have been that when using the FormHelper text_field method, calling the `tag_list` method did not return a comma separated string. So we explicitly set the value to be the stringified array, as suggested in https://github.com/mbleigh/acts-as-taggable-on/wiki/Add-a-tag-field-to-forms-in-Rails--4.2-and-above (thanks @h-lame!)

We update the step definition for "I should see the tags..." so that the order of tags is irrelevant. By default, using `acts_as_taggable` on a model does not guarantee the exact ordering of tags on save.