Remove whitespace from tag_list in cucumber step
The test was failing in CI because when passing a comma-separated string
and turning it into an array, the leading white spaces would not be removed
and therefore elements would not be matched correctly.

```
expected collection contained:  [" fi", " fo", "fee"]
actual collection contained:    [" fee", " fi", "fo"]
```

This commit removes leading and trailing whitespace for the elements.