Handle gds_api responses as hashes not openstructs
In the previous commit we turned on the new `hash_response_for_requests`
config option and in this commit we pay the price.  The code works in two
modes; direct from the API or reading from yaml.  Making the YAML version
symmetric with the new api adapter behaviour is simple enough - we stop
pushing the YAML loaded data through the `GdsApi::Response.build_ostruct_recursively`
method and just use the raw YAML loaded data (which is an array of hashes
already - the same as the API response).  We do also have to change the
YAML to use string keys, as the API response will not return symbol keys.