Add govuk-lint to project and modernise jenkins.sh
We add linting to the default rake task as the final task to run. This
means that we'll run lint twice on jenkins (once because it's in
jenkins.sh explicitly and once from the default rake task).  Instead we
add a rake nolint that runs the default rake task without lint, and make
jenkins.sh to run taht instead of the default rake task on its own.