Correct intention of User#can_edit_site?
We had boolean logic declared as:

    a || b || c || d && e

Our intention was:

    (a || b || c || d) && e

But precedence rules in ruby actually meant it was interpreted as:

    a || b || c || (d && e)

In this example `a`-`d` are conditions based on the user roles and
organisation memeberships, and `e` is a condition based purely on the
site and it's `global_type`.  The result of our logic being expressed
wrongly is that some people were allowed to edit global sites when
the intention was that no-one should be able to edit global sites.

We spotted this because rubocop wanted to indent the expression oddly
and this led us to work out the true intention of the code.  To fix we
add some specs and clarify the expression to make it true to our intent.