Use same global site check in batch controllers as mapping controller
You can't edit the mappings for a global site so the mappings controller
will redirect back to the site root with an appropriate message.  The two
batch controllers (import batch controller, bulk add batches controller)
allow similar things, but do not perform this check.  They do perform a
standard check to see if the user can edit, but this redirects to the
site mapping path, which in turn will perform the global check and
redirect to the site path.  So we get the correct result, but it costs
an extra redirect and isn't as clear an intent as it could be.

We extract a concern with the mappings controller behaviour and make
use of it all three controllers.

Note we've made this concern one you must explicitly opt into rather
than adding it to all controllers as it's only used in 3 places.