Change column types on mappings_batch_entries to match mappings
In particular because archive_url was too small for National Archives URLs
which are usually long. Batch entries are the precursor to mappings so their
column types should match where possible, so that users can create mappings
via batches with the same constraints as when doing them individually.

Also use a long archive_url in a test which fails without this migration.