Suppress javascript errors in Poltergeist
We've got intermittent-ish test failures around the charts js in Transition
again, which get in the way of us making even small unrelated changes to the
app. This time it's variations on this in the hits_summary feature.

We're disabling raising of javascript errors in Poltergeist so that we can get
builds running again.  We've got a [card for investigating the issue further][1]
in our backlog.

Trello story: https://trello.com/c/nvjibWTy/633-transition-test-failures

[1]: https://trello.com/c/kaQn2aYB/528-transition-test-failures