Only expose hosts with a CNAME or IP address
This is intended to reduce the number of hosts registered with Fastly
through the configure_bouncer script in cdn-configs. There are
currently 3476 domains being registered, and the current limit is
3500.

The rationale is that if Transition cannot find a CNAME or IP address
coresponding to the host, then it is not necessary to register it with
Fastly.

In the Transition database, roughly 1365 hosts have a CNAME and 559
have an IP address. There are 1552 hosts without a value for either of
these. It's these 1552 hosts that this change will stop Transition
from exposing, which is expected to cause the configure_bouncer
script to remove these domains from Fastly.

The expectation here is that this won't cause any problems for
existing or new sites, as some CNAME or IP address is required for
requests to make it to Fastly. For new sites, it is expected that the
hostname will also exist prior to traffic being handled through
Fastly, so Transition should still expose these hosts.