Remove code to ingest Google Analytics
Confirmed by @jennyd that this code isn't being used:

> Not used - it was an early possible way of getting traffic data into
the app. The plan was that maybe depts would give us access to their GA
accounts, but we never did that - we got their server logs instead.

Removing it will make it easier to get rid of alphagov-deployment, as
that still contains Google credentials for this feature.