Update a test for a host which isn't redirected by GDS
When we originally added this test back in 2013 [1], we were only supporting
transitions via CNAME, not A records. So this test represented the
straightforward negative case then.

Now the situation is more complex, and transitions via various sets of IP
addresses are supported as well as CNAMEs. There are other tests already for
most of those cases, so update this test to cover the case of no relevant DNS
records being found.

[1]: https://github.com/alphagov/transition/commit/e5302d968bba166ee02fbe5920204dec44752421