(fix) Switch to port 3000 to enable SSO callbacks
Using port 9292 is the Puma default but is unconventional for Rails. I had recently overlooked this as we already have staging containers running that were probably using this port. Changing the port again would require work.

When trying to sign in I noticed another good reason to keep with conventions. The callback URL has been set to allow localhost:3000 not localhost:9292, and an error is thrown.

I think the right thing to do is to switch back to the exclusive use of port 3000 in all cases.