Remove `track_associations = false` in PaperTrail initializer
There was a deprecation warning in the CI output for this:

```
Association Tracking for PaperTrail has been extracted to a separate
gem. To use it, please add `paper_trail-association_tracking` to your
Gemfile. If you don't use it (most people don't, that's the default) and
you set `track_associations = false` somewhere (probably a rails
initializer) you can remove that line now.
```