Reach into paper trail to display originator
The WhitelistedHost model does not contain an `originator` field.
Instead this is exposed by `paper_trail` (a gem that provides version
history on a model) which the model has referenced.

Update the view so we refer to `originator` through `paper_trail` rather
than directly on the model. This should fix the error we are seeing
"ActionView::Template::Error: undefined method `originator'".