Rubocop: make changes to cops
There are a number of linting errors for migration files from
before 2020, which cannot be fixed automatically. The time
and effort to fix, and the risk of an incorrect fix, makes it
not worth doing.

Also exclude the db/schema.rb file as this is auto generated

Ignore style on bin/setup and bin/update because it was causing errors that couldn't be fixed