Include ActiveSupport::LoggerSilence
After upgrading activerecord-session_store we hit a problem with our
logging (see
https://github.com/rails/activerecord-session_store/issues/176 for more
details).

The logger provided to us by govuk_app_config doesn't have support for
the silence method, so we can add it in place here. If we find that
we're needing to this to lots of our applications, we should move it
into govuk_app_config.