Only require necessary rails frameworks
After upgrading to rails 6.1 we saw an error occur:
```
ERROR -- : Couldn't find Active Storage configuration
```
Despite the fact that we aren't using Active Storage

We have traced this to be related to the fact that we require all of the
rails frameworks even though we don't use them all.
It is mentioned in this issue: https://github.com/rails/rails/pull/39498

To prevent this happening in future we are requiring only the frameworks
that the application uses.