Bump papertrail
After upgrading to rails 6.1 we were getting the following error:
```
PaperTrail 11.0.0 is not compatible with ActiveRecord 6.1.4.1. We allow PT
contributors to install incompatible versions of ActiveRecord, and this
warning can be silenced with an environment variable, but this is a bad
idea for normal use. Please install a compatible version of ActiveRecord
instead (>= 5.2, < 6.1). Please see the discussion in paper_trail/compatibility.rb
for details.
```

This compatibility issue was resolved in a release by paper_trail.
Therefore upgrading the gem fixes the problem.