Conservative update of sentry-sidekiq 5.6.0
This removes the extra updating of gems that were also updated by
sentry-sidekiq. This was achieved by running:

```
git checkout HEAD~1 -- Gemfile.lock
bundle update --conservative sentry-ruby sentry-rails sentry-sidekiq
```

The motivation for this is to revert the major version update of
paper_trail. This update cannot be applied automatically because we use
the YAML serializer that has a backwards incompatible change in
paper_trail 13 [1]. I thought it was easier to revert this distinct
dependabot update and instead deal with paper_trail later when there is
a distinct depenabot PR for it.

I'm expecting we're going to have to set-up something like [2].

[1]: https://github.com/paper-trail-gem/paper_trail/blob/master/doc/pt_13_yaml_safe_load.md
[2]: https://github.com/paper-trail-gem/paper_trail/blob/v12.3.0/README.md#convert-existing-yaml-data-to-json