Remove tools partial
Following the removal of the side-by-side browser I looked at the other
tools here and determined 2 of the remaining 4 were broken. The 2 that
are broken are the "Editing a mapping for a URL" (legacy domain name)
and "Convert a page to Markdown" (3rd party resource 404s).

Given these are so poorly maintained it seemed sensible to remove all
the tools and then we don't need to worry about the others breaking.

The expectation is that since 2015 there has been very low usage of these
features.