Add `Site manager` permission
We are in the process of archiving Transition Config, which requires that we
add various bits of functionality to Transition.

In order to test these safely, we have added a new `Site manager` permission to
Signon.

This adds the permission to the user model, as well as some test helpers.

Co-authored-by: CristinaRO <579522+CristinaRO@users.noreply.github.com>