Add site form object
We are aiming to archive Transition Config which is currently used to create
sites in Transition.

This adds a form object for the creation of a site. 

It's worth noting that the process currently moves existing hosts and aka hosts
to the new site if they already exist (see
`lib/transition/import/site_yaml_file.rb`). We aren't yet sure whether we want
to create a new UI for that functionality, or to use this form. For now, this
form does not have that functionality.