Add PaperTrail to the Site model
Requires including `PaperTrail::Rails::Controller` in the site controller.

I've also had to fix the order of before_actions in the application
controller, as the user name was not available to PaperTrail in the old
order.

The controller specs didn't pick up the ordering bug, likely due to the
way that authentication mocking is done. However, request specs did
reproduce the issue, and therefore proved the fix.