Update the new site form to use GOV.UK styles
Use the Publishing Components for the new site form.

This introduces a form helper to help us work with the Publishing Components in
a consistent way.

The extra organisations field does not use the Publishing Components, we
currently use the `multiple` attribute to select multiple organisations and
this can't be chosen with the gem. As a further refinement we might want to
consider replacing this with the autocomplete multiselect component or similar
(alphagov/accessible-autocomplete-multiselect).