Move delete site confirmation page to the Design System
Now that we have introduced a layout for the Design System, we should ensure
that we are using it when adding new pages.

The most appropriate way of showing the `The confirmation did not match`
message while following the Design System seems to be to use form validation,
so this also introduces a form object to back the delete form. This has the
additional benefit of moving the deletion logic out of the controller.