Add form for editing a transitioned site
This was previously done by editing the information in YAML files in the
`transition-config` repo.

Adding a user interface so changes can be made without developer
intervention.

Any related hosts are unchanged if they already exist (based on hostname).  If
a host has been removed from a site, it's record is destroyed. If a hostname is
new, a new record is created.

Validation will soon be added to prevent a host being created with a hostname
that is already in use.

Co-authored-by: BruceBolt <bruce.bolt@digital.cabinet-office.gov.uk>