Validate that hostnames are unique
The current import process will move a host to a new site if a site with that
same host already exists (documented in
`spec/lib/transition/import/site_yaml_file_spec.rb:177`).

We think this will not happen frequently, so have decided to disallow adding or
updating a host where the hostname is not unique.

All hostnames in the database are currently unique, so this validation can run
on create and update without invalidating any current records.

Note that this executes 2 SQL queries, whereas we could perform this in 1 if
using custom validation, however I think using the built in validation in
clearer for this use case.