Add rake task to remove a host
There may be times that a host needs to be removed. This is currently
not possible (although will be when we make the edit site UI available
to users).

Therefore adding a rake task to allow us to remove hosts when they are
no longer needed.

The corresponding change will also need to be made in
`transition-config` to stop the host being reimported again.

Co-authored-by: Aga Dufrat <aga.dufrat@digital.cabinet-office.gov.uk>
Co-authored-by: Mahmud Hussain <mahmud.hussain@digital.cabinet-office.gov.uk>